Backend/requirements.txt
alabaster==0.7.10
appnope==0.1.0
argon2-cffi==16.3.0
attrs==17.3.0
awesome-slugify==1.6.5
Babel==2.5.1
backports.shutil-get-terminal-size==1.0.0
boto3==1.4.7
botocore==1.7.48
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
contextlib2==0.5.5
coreapi==2.3.3
coreschema==0.0.4
coverage==4.4.2
decorator==4.1.2
Django==1.10.8
django-allauth==0.34.0
django-activity-stream==0.6.4
django-anymail==1.2
django-coverage-plugin==1.5.0
django-crispy-forms==1.7.0
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-extensions==1.9.7
django-model-utils==3.0.0
django-redis==4.8.0
django-rest-framework-nested==0.0.1
django-rest-swagger==2.1.2
django-storages==1.6.5
django-test-plus==1.0.20
djangorestframework==3.7.3
python-memcached
docutils==0.14
enum34==1.1.6
factory-boy==2.9.2
Faker==0.8.7
funcsigs==1.0.2
futures==3.2.0
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
idna==2.6
imagesize==0.7.1
ipaddress==1.0.18
ipdb==0.10.3
ipython==5.5.0
ipython-genutils==0.2.0
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
oauthlib==2.0.6
olefile==0.44
openapi-codec==1.3.2
pathlib2==2.3.0
pexpect==4.3.0
pickleshare==0.7.4
Pillow==4.3.0
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
py==1.5.2
pycparser==2.18
Pygments==2.2.0
pytest==3.3.0
pytest-django==3.1.2
pytest-sugar==0.9.0
python-dateutil==2.6.1
python-openid==2.2.5
pytz==2017.3
raven==6.3.0
redis==2.10.6
regex==2017.11.9
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.12
scandir==1.6
simplegeneric==0.8.1
simplejson==3.13.2
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.6.5
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.4
termcolor==1.1.0
text-unidecode==1.1
traitlets==4.3.2
typing==3.6.2
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
whitenoise==3.3.1
